+++ lib/require_profiler/plugins.rb
@@ -6,6 +6,7 @@ module RequireProfiler
       def register_reporter(reporter)
         HTTPPlugin.new(reporter).activate! unless ENV["REQUIRE_PROFILER_HTTP"] == "false"
         YAMLPlugin.new(reporter).activate! unless ENV["REQUIRE_PROFILER_YAML"] == "false"
+        RailsPlugin.new(reporter).activate! unless ENV["REQUIRE_PROFILER_RAILS"] == "false"
       end
     end
 
@@ -19,5 +20,6 @@ module RequireProfiler
 
     autoload :HTTPPlugin, "require_profiler/plugins/http_plugin"
     autoload :YAMLPlugin, "require_profiler/plugins/yaml_plugin"
+    autoload :RailsPlugin, "require_profiler/plugins/rails_plugin"
   end
 end